+++ src/main/java/org/folio/entlinks/domain/entity/projection/AuthorityId.java
@@ -1,0 +1,8 @@
+package org.folio.entlinks.domain.entity.projection;
+
+import java.util.UUID;
+
+public interface AuthorityId {
+
+  UUID getId();
+}