+++ src/main/java/org/folio/entlinks/domain/entity/projection/AuthorityIdDto.java
@@ -1,0 +1,6 @@
+package org.folio.entlinks.domain.entity.projection;
+
+import java.util.UUID;
+
+public record AuthorityIdDto(UUID id) {
+}