+++ src/main/java/org/folio/entlinks/domain/repository/AuthorityRepository.java
@@ -5,9 +5,11 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 import org.folio.entlinks.domain.entity.Authority;
+import org.folio.entlinks.domain.entity.projection.AuthorityId;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
@@ -16,6 +18,9 @@ public interface AuthorityRepository extends JpaRepository<Authority, UUID>, Aut
   List<Authority> findByNaturalIdInAndDeletedFalse(Collection<String> naturalIds);
 
   Page<Authority> findAllByDeletedFalse(Pageable pageable);
+
+  @Query("select a.id as id from Authority a where a.deleted = false")
+  Page<AuthorityId> findAllIdsByDeletedFalse(Pageable pageable);
 
   Optional<Authority> findByIdAndDeletedFalse(UUID id);